SysLink: RCM - Recipient structure missing elem
The Recipient structure definition in RcmClient.c was missing the member
List_Elem elem.  This member allows the Recipient object to be used in a List.
In the context of RcmClient, it would be used to create a recipient list for
other threads to signal, if no message for the current recipient was found.
But since the struct was missing elem, it could not be correctly used with List.
Previously this did not cause problems because we did not test multiple threads
on one client, so we did not have to switch between recipients with blocking
RCM calls.

Signed-off-by: Simon Que <sque@ti.com>