SysLink: SysMgr - Decouple SysMgr from ProcMgr module in kernel
This patch adds the functionality to invoke the
SysMgr start, stop, and load callback. This patch
is intended to de-couple SysMgr calls from ProcMgr
in kernel module.

The SysMgr calls are now called from ProcMgr functions
in user-space. The long term solution should be to
even de-couple this from User-space.

Signed-off-by: Hari Kanigeri <h-kanigeri2@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>