SysLink: Build - Compilation warnings during autoconf build removed.
The README file updated with the tools details and the option to
enable DEBUG option.
Few options added in the build_directions.sh file to prompt the user to
choose whether to build bridge or syslink code.
Configure.ac file changed to disable static build in the syslink code.
Debug option by default is OFF now.

Signed-off-by: Imtiaz Ahmed <x0116819@ti.com>