SysLink: Utils - MemoryOS setup in sysmgr get config
MemoryOS module is initialized in SysMgr_getConfig
function by calling UsrUtilsDrv_setup function. This
is required to initialize in this function as the
functions in SyMgr_getConfig may call MemoryOS module
functions and it is required for this module to be
setup.

This fixes "MemoryOS_copy Module was not initialized"
error.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Hari Kanigeri <h-kanigeri2@ti.com>