SysLink: ProcMgr - Turn off a ProcMgr_delete failure trace
ProcMgr_delete API is called as part of ProcMgr_destroy from
SysMgr_destroy. As such, this API is always called from an
application calling SysMgr_destroy. This results in a misleading
trace statement print when the application is run alongside the
Syslink daemon. This trace is therefore converted from a
GT_setFailureReason to a regular GT_trace to minimize confusion
to end-users.

Signed-off-by: Suman Anna <s-anna@ti.com>