SysLink: ProcMgr - Handle invalid loader return value
ProcMgr_load was checking the status of load_executable by checking prog_handle.
If it was null then it was indicative of a failure.  However, if SysM3 load
succeeded and AppM3 image was not found, prog_handle would still have the old
valid value.  load_executable returns -1 in that case.  So now ProcMgr_load also
checks for the return value of load_executable.

Signed-off-by: Simon Que <sque@ti.com>