SysLink: Daemons - Fixed signal handling and cleanup
This patch fixes two bugs.

First, it removes the call to exit() in the signal handler.  Calling exit()
caused the problem of shutting the application down abruptly, without giving
the daemon thread a chance to finish cleanup.  With this change, the daemon
can finish cleanup and exit by returning from main().

Second, it removes an extra AppM3 usage flag (Bool appM3Client) in ipc_setup()
that was masking the global flag.  The variable should be global because it is
being used by multiple functions.  The local flag prevented ipc_setup() from
communicating the AppM3 status to the other functions, interfering with the
cleanup process.  Previously undetected because the exit handler was not
working so it never reached the cleanup phase.

Signed-off-by: Simon Que <sque@ti.com>