SysLink: IPC - SharedRegion migrated to SysLink 2.0
SharedRegion module is synched up with SysLink-2.0 version.
This version incorporates changes for improved ease of
programming for the user.

The new SharedRegion module is a drastically different module
compared to before. It supports heap creation in-built into
the regions provided by the new HeapMemMP module. This is used
for automatic gate creation and other shared object creation.

The following are the main changes specific to this module:
1. Old API is deprecated. It retains only a couple of
   old API.
2. A whole set of new API are provided. The list includes
   SharedRegion_attach, SharedRegion_detach, SharedRegion_start,
   SharedRegion_stop, SharedRegion_reserveMemory, SharedRegion_setEntry,
   SharedRegion _clearEntry and SharedRegion_clearReservedMemory
   among others.

The following are the changes in line to the common changes to all
SysLink-2.0 IPC modules.
1. Module headers reflect the new common header files in
   include/ti/ipc/ folder.
2. Module statuses are converted accordingly (SHAREDREGION_E_error changed
   to SharedRegion_E_error & SHAREDREGION_success to SharedRegion_S_success)
3. A new _SharedRegion.h internal header file is added.
4. A SharedRegion_module variable is introduced to mimic the BIOS behavior.
5. BSD License adoption.

Signed-off-by: Suman Anna <s-anna@ti.com>