SysLink: IPC - GateMP module added from SysLink 2.0
A new module GateMP is added. This module is added for SysLink-2.0
to improve ease of programming for the end-user.

This module abstracts out all the multiprocessor gates, and is
the primary interface for application users wanting multi-processor
protection. Users no longer would have to deal with different gate
API like GatePeterson or GateHWSpinLock.

Signed-off-by: Suman Anna <s-anna@ti.com>