SysLink: IPC - HeapMemMP module added from SysLink 2.0
SysLink 2.0 contains a new module, HeapMemMP.  This patch
adds that module to our SysLink code base.

HeapMemMP differs from HeapBuf/HeapBufMP in that it allows
allocation for variable sizes.  It is designed for use when the
size of allocation is not known beforehand or when there is
a large variation in expected alloc sizes.

HeapMemMP includes the various API/module changes that are
common to all modules in SysLink 2.0 such as:
- Status definitions follow the format HeapMemMP_E_<status>.
- The HeapMemMP_Params_init function does not take a handle
  argument.
- HeapMemMP_open opens an existing HeapMemMP by name only.
- HeapMemMP_openByAddr opens an existing HeapMemMP based on
  shared memory address.
- BSD License adoption.

Also, we have retained the use of an enum for IOCTL codes
in HeapMemMPDrvDefs.h.

Signed-off-by: Simon Que <sque@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>