SysLink: RCM - Post to semaphore twice for Android workaround
Android's sem_destroy throws an EBUSY error on a semaphore that is has
no threads waiting on it in RcmServer.  To resolve this, RcmServer posts
to it twice. This allows the cleanup to proceed.  This should not break
RcmServer on Linux because the semaphore is used only once (one pair of
pend/post).

The second semaphore post is added conditionally, and only gets executed
for Android.

Signed-off-by: Simon Que <sque@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>