SysLink: Notify - NotifyDriverShm module migrated to SysLink-2.0
NotifyDriverShm module is synched up for SysLink-2.0. This version incorporates
changes for improved ease of programming for the user. This module has been
removed completely from user-space. This module is solely present on the
kernel-side and is configured as part of the SysMgr. This module is not
used directly anymore, and is abstracted out by the Notify module.

Signed-off-by: Suman Anna <s-anna@ti.com>