SysLink: SysMgr - SysMgr module converted to Ipc module
SysMgr module is synched up for SysLink-2.0 version. This version
incorporates changes for improved ease of programming for the user.

SysMgr module is now renamed to a new Ipc module to reflect the name
properly on both MPU & Slave processors. The functionality of SysMgr
remains mostly the same.

The following are the main changes specific to this module.
1. All 'SysMgr' files have been renamed to the corresponding 'Ipc' files.
   information is captured in _Ipc.h and a common header file between
   MPU & Ducati, ti/ipc/Ipc.h

TODO:
Ipc.c
1. Retained Ipc_loadcallback, Ipc_startcallback and Ipc_stopcallbacks.
   Need to get rid of these and move this logic into Ipc_control. May
   need to add the check for procId before sending down the ioctl. Also
   need MultiProc usage rather than #defined PROC_ID limit check.
2. Define the commands appropriately.
3. Rename the files to Ipc.
4. Move it within Ipc folder.

IpcDrvDefs.c
1. Remove the custom defines for load, start and stop callbacks.
2. Modify the DrvCmds structure to reflect the above change.
4. Need the IPC_BASECMD change in IpcCmdBase.h

IpcUsr.h
1. Remove the load, start and stop callback definitions.

Signed-off-by: Suman Anna <s-anna@ti.com>