SysLink: Samples - Updated SharedRegion sample for SysLink 2.0
SharedRegion sample has been updated to work with the new SysLink-2.0 IPC
version. The new sample uses Ipc/ProcMgr and loads the Notify Reroute sample
as part of the setup. It also initializes SR entries that are independent of
the default region(s) created by IPC setup.  It uses these regions for testing.

The sample is cleaned up to remove all the obsolete non-SysMgr variant code.
This is no longer needed with the latest IPC code.

Signed-off-by: Simon Que <sque@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>