SysLink: Daemons - Create default HeapBufMPs for use by DOMX
Two default HeapBufMPs are created within the SysLink daemon with
the new SysLink code. There are no heaps that are created by default
within SysMgr, and the daemon creates and registers the heaps to be used
by all other applications. Heap0 is used by TILER and Heap1 will be
registered for usage by DOMX Clients.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>