SysLink: IPC - Make SysLink loadable modules for SysLink 2.0
In order to make SysLink into loadable/unloadable modules, the
notify, notify_ducatidriver, and syslink_ipc modules were combined
into one module.  This needed to be done because of the circular
dependencies that exist between the three modules.  Also, the
platform_mem module had to be moved to its own module.

There are corresponding kernel-side changes needed with this patch.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>