SysLink: ProcMgr - Handle Resets properly in ProcMMU
Added reset for RST3 to ProcMMU_close so that resets are in proper state for
next run.  Also moved ProcMMU_close to be called after CMD_PROCMGR_STOP.

In ProcMMU_open, if the resets are found to not be in the proper state, it
sends a reset for RST1, RST2 and RST3, but doesn't wait for it to complete
before continuing.  ProcMMU_open should wait for the reset completion before
continuing or the subsequent de-asserts may not work properly.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>