SysLink: Loader - Fix munmap calls
The munmap call in DLIF_write was performing an munmap on an
un-initialized variable and the munmap call in DLIF_release
was calling munmap on the wrong address.  This patch updates
the code so that the proper address is passed to both munmap
calls.  It also updates munmap to be Memory_unmap, since
Memory_map is used for mapping the memory.

Amended the patch to fix some compiler warnings.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>