SysLink: ProcMgr - Fix the cache function direction argument
The Cache flush and invalidate argument has to match for the buffer.
The Kernel internally stores the flush direction when a cache flush
is called, and when invalidate is called on the same buffer it checks
for the previous flush function direction and if they don't match the
cache invalidate function fails.

Signed-off-by: Hari Kanigeri <h-kanigeri2@ti.com>