SysLink: ProcMgr - Move ProcMMU_open call to ProcMgr Open
The call to open ProcMMU is moved to ProcMgr_open. Without this,
the handle to IOMMU is invalid since previously it was opened only
during the Load time.

So one would see the DMM test cases fail if the Application is built
with no loader option, or run with the SysLink Daemon.

Signed-off-by: Hari Kanigeri <h-kanigeri2@ti.com>