SysLink: ProcMgr - Fix memory mapping leaks
ProcMgr_open and ProcMgr_attach API are used in mapping the processor
related entries into the user-space. The mapping is done in one of the
API based on whether the kernel proc_mgr module has been attached or not
previously. In the case of the application running alongside the daemon,
the mappings are done in ProcMgr_open, but ProcMgr_attach resets the
memEntries count thereby leading to a memory unmap leak in ProcMgr_close.
This patch fixes this defect by introducing separate counters to keep
track of entries opened during _open or _attach and use them to do the
unmapping in the respective _close or _detach functions.

Signed-off-by: Suman Anna <s-anna@ti.com>