SysLink: ProcMgr - Add non-hwmod support for slave resets
Add the Ducati/Tesla reset and unreset sequences temporarily in
userspace ProcMMU module for Android as hwmod functionality is
not yet supported in the current Android kernel.

Group the reset/unreset sequence code under android specific flag,
as these changes are not required for the Ubuntu release stream.

NOTE: This is a temporary patch and needs to be reverted once hwmod
is supported for Android.

Signed-off-by: Shivananda Hebbar <x0hebbar@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>