SysLink: ProcMgr - Incorrect proc id passed to ProcMMU close
This patch fixes an issue where an invalid procId is passed
to ProcMMU_close from the ProcMgr_close function.  When
ProcMMU_close is called, a pointer that has been freed is
dereferenced to get the procId.  This results in the procId
being 0 when ProcMMU_close is called.

This patch saves the procId before freeing the pointer and
uses the saved procId value to call ProcMMU_close so that
the correct procId is passed.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>