SysLink: Samples - Start Notify sample from event 4
Update the Notify sample so that the sample starts the registration
from event 4 instead of 3. This is needed as event 3 is reserved
for IPC/SLPM. A similar change on BIOS-side will require this patch.
The BIOS-side change is required though so that the BIOS application
does not register for event 3.

Signed-off-by: Suman Anna <s-anna@ti.com>