SysLink: ProcMgr - Handle Tiler address map in userspace
Handle the Tiler address space in userspace until Tiler phase 2.0
is available. The users are expected to call Tiler's VirtToPhys on
the memory returned by MemMgr Alloc, and provide this address to
SyslinkMemUtils_map function.

When the Phase 2.0 is enabled, the SyslinkMemUtils_map should call
ProcMgr_map function with Anonymous flag.

Signed-off-by: Hari Kanigeri <h-kanigeri2@ti.com>