SysLink: Daemons - Add delay to syslink daemon for recovery
This patch provides a temporary fix for a timing issue where
the iommu may not be recovered before the daemon tries to
re-open the iommu handle.  This patch adds a small delay
before opening the iommu handle in order to give the iommu
sufficient time to recover.  This is only temporary until
a more robust solution is available.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>