SysLink: Samples - Modify mmufault_listener to support more events
Add the capability for listening more events in
mmufault_listener sample and rename to a more
generic name.

NOTE: This samples is useful as an example how the APPs can
register for events like MMUFAULT and PROC_STOP. Because in
those cases the APPs need to close their handle and wait until
the daemon is restored.

Signed-off-by: Fernando Guzman Lugo <x0095840@ti.com>