SysLink: Daemons - Crash Dump utility cleanup
Assignment to ExcContextBuffer is done using a typecast rather than
manually walking strucuture memory. The case statement for
determining type of thread on which exception occured did not
have a 'default' catch; it now does. Added a typedef for the max
size of ExcStackBuffer rather than using a magic number.

Signed-off-by: Wajahat Khan <w-khan@ti.com>