SysLink: Samples - Adding delay to simulate system suspend in resources sample
The resources test simulate System suspend using ioctl functions that
sends the suspend-resume messages to Ducati. Suspend message will trigger
hibernation on Ducati which expects to be rested by A9. In System suspend,
Ducati is indeed reset but not in the ioctl path. If the resume message is
sent immediately after the Suspend path, this will interrupt Ducati in the
middle of the saving context routine which is a high priority Swi. The
resume message will be enqueued to be attended later in a lower Swi. The
high priority saving context swi continues then and finishes calling WFI,
so Ducati can only attend other tasks until reset or a mbox message arrives.
Since the resume message has been already received but enqueued in a low
priority Swi, this is not attended until another mbox message arrives or
Ducati is reset.

In the resources test Ducati is not reset, so if we add a delay we assure
the resume message is sent when Ducati is already in WFI and not in the
middle of the high priority Saving context Swi avoiding the above described
issue.

Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Miguel Vadillo <vadillo@ti.com>
Signed-off-by: Paul Hunt <hunt@ti.com>