SysLink: RCM - Enhance current RCM design
The following two major RCM design enhancements are introduced
with this patch:
1. RCM Server Worker Pool design
The current RCM design has only a single server thread which
processes all the incoming messages sequentially. The RCM Server
Worker Thread Pool design creates a RCM Server with multiple pools
of worker thread. This allows users to prioritize each call to
RCM server. If you create an RcmServer instance using just the
default values, the default pool will not have any threads. All
messages sent to the default pool will be processed by the server
thread function (in-band processing). The RCM server allows for
creation static worker thread pools during create phase. Dynamic
creation of worker thread pools is not supported at this time.

On the client side, RcmClient_alloc() will initialize the message
with "poolId = RcmClient_DEFAULTPOOLID" (this is default value)
thus addressing the message to the default pool. The user can
change the poolId to send the message to a particular pool. When
the server thread function receives this message, it checks
the poolId to route the message to correct worker thread pool
or process the message itself (in-band processing). Additionally,
the client can acquire and assign a jobId to the RCM messages.
RCM messages sent with the same job stream id are guaranteed to
be processed serially. The default value for job stream is set to
RcmClient_DISCRETEJOBID, and these are processed by one of the
respective pool's worker threads.

2. Error Checking on uni-directional RCM messages
The RCM Client now has an error queue to which all the failed
RcmClient_execCmd (formerly RcmClient_execNoReply) RCM messages
are returned to. The user can check for these messages using the
new RcmClient_checkForError API. The RcmClient_checkForError API
returns the same statuses as the regular RcmClient_exec API.

The patch also includes SysLink 2.0 userspace code compliant
formatting changes.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Shahid Akhtar <sakhtar@ti.com>