SysLink: RCM - Use MessageQ_unblock in RcmServer
The RcmServer_delete call is modified to call MessageQ_unblock and the
workaround in the server thread function where MessageQ_get() is called
with a 1 second timeout value has been removed. MessageQ_get() is called
with the expected FOREVER timeout value, and the MessageQ_unblock helps
in unlocking the ioctl call in the kernel.

This patch should resolve the 1 second delay problem while shutting down
the server.

Signed-off-by: Shahid Akhtar <sakhtar@ti.com>