SysLink: Daemons - Add thread join protection in SysLink Daemon
MMU fault thread is exiting by posting the semaphore, but before
the kernel is cleaning up the thread resources the thread is
recreated with the same handle since the kernel didn't delete
the thread yet. After several iterations, it will return a
failure EAGAIN when trying to create the thread.

So apparently after several iterations, system is running out
of resources or PTHREAD_THREADS_MAX is reached. To avoid this,
the thread handle is set to NULL before passing to pthread_create,
pthread_join was added to ensure the termination of the already
created thread before creating a new one, also now a check for
pthread_create was included.

Signed-off-by: Miguel Vadillo <vadillo@ti.com>