SysLink: Daemons - Avoid thread join if not recovering
If the syslink_daemon was killed or exited by any other
way but the mmu_fault_handler it was hanging because
it was waiting in the pthread_join for mmu_fault_handler
thread termination but this thread was never reaching the
end.

Add a check to call pthread_join just in the restart case
and in the signal_handler callback kill the mmufault thread.

Signed-off-by: Miguel Vadillo <vadillo@ti.com>