SysLink: ProcMgr - Change rproc and iovmm device permissions to RDONLY
Remoteproc and iovmm devices are opened with RDWR permissions, and this
requires that the device files be given these broad permissions so that
a variety of applications can use SysLink. These permissions are too
broad and needs to be limited. The iommu and rproc drivers are opened
now only with O_RDONLY permissions that allows the limited file
permissions on the device files.

Signed-off-by: Shivananda Hebbar <x0hebbar@ti.com>