SysLink: Samples - Update memmgrserver sample
The alloc and free functions are provided as library functions in
SyslinkMemUtils, so we don't need the equivalent functions in
the sample. The sample registers the library provided functions,
SysLinkMemUtils_alloc and SysLinkMemUtils_free instead. The TilerMgr
calls are also removed. The sample has been cleaned up to remove
other unused functions.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>