SysLink: ProcMgr - Return MemMgr status in SysLinkMemUtils_free
The status from the TILER MemMgr_free is ignored currently in the
SysLinkMemUtils_free function. This is fixed and the status of the
MemMgr_free is seen on the BIOS-side code. The SysLinkMemUtils_unmap
error status is printed out, but is not used to rule out calling the
MemMgr_free function.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>