SysLink: Daemons - Enhance to recover for PROC_ERROR on M3 cores
The SysLink daemon is enhanced to listen for system errors on the
SysM3 & AppM3 processors, and reload them in case of an error. It
uses Devh interface to register for syserror events from the devh
devices in kernel.

The previous MMU Fault thread is converted into a SysM3 listener
thread, and a new thread is created to listen for error events on
AppM3.

Signed-off-by: Wajahat Khan <w-khan@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>