SysLink: RCM - Check status returned by MessageQ_unblock
The RcmServer_delete API is fixed to check the status of the
MessageQ_unblock API, before calling the pthread_join on the
main server thread. Without checking this status, it is not
guaranteed that the server thread has been unlocked successfully
resulting in the RcmServer_delete API getting blocked and
thereby hanging the application.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>