SysLink: Daemons - Enhance to recover from M3 WatchDog events
The SysLink daemon is enhanced to listen for WatchDog events on the
M3 processors, and reload the M3 processors in case of an error. Some
heavy code formatting cleanup changes are also incorporated.

The current kernel WatchDog logic is integrated with the SysLink
IPU_PM hibernation logic, and cannot distinguish whether a WatchDog
has occured on SysM3 or AppM3, and would notify both in case of one.
However, the daemon code is listening to both the cores still, to
account for the case that AppM3 may not be loaded always.

The pthread termination sequence has been modified to clean up more
graciously.  The pthread_kill calls have been removed from the signal
handler, and moved to the main thread. The signal handler posts the
semaphore that would be used as a trigger to kill the threads from
main thread.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Hari Kanigeri <h-kanigeri2@ti.com>