SysLink: Utils - Make LOG_TAG customizable
These changes allow LOG_TAG to be modified in any executable.
The LOG_TAG macro variable sets the tag seen when printing
traces to the Android logging system (logcat).  When SysLink
is built under Android, Osal_printf() output is redirected to
Android logging functions, i.e. LOGD, if the macro
SYSLINK_LOGCAT_ENABLE is defined at the linking stage. These
functions use LOG_TAG to prefix an identifier to the output.

To set the tag for a file:
   #include <OsalPrint.h>
   #if defined (HAVE_ANDROID_OS)
   #define LOG_TAG "something"
   #endif

Any Osal_printf calls following the #define will be prefixed
with "something".

The SYSLINK_DIRECT_LOGD macro, if defined when compiling
any sources, will replace any Osal_printf calls with LOGD.
When this macro is not defined, Osal_printf() is still called,
but is now configurable.  The Osal_configPrint() function
can be used to redirect the traces to LogCat selectively.
To redirect these traces, do:

   Osal_printConfig (Osal_printToLogCat);
   ... other function calls, code ...
   Osal_printConfig (Osal_printToConsole);

Default behavior is to output traces to the console.

For non-Android builds the output from Osal_printf() remains
directed through stdout.

Signed-off-by: Cris Jansson <cjansson@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>