SysLink: Daemons - Prevent multiples instances.
This is a preliminary implementation of a feature to
prevent running mulitple instances of the Syslink daemon.
The detection is done by parsing the procfs to see if
there is more than one process with the name of the
Syslink daemon.  If it finds another instance running,
it will exit immediately.

This patch should be reverted once the PROCMGR state
machine is implemented.

Signed-off-by: Subin K G <subin.kg@ti.com>