SysLink: ProcMgr - Remove memory leak in ProcMgr_waitForMultipleEvents
After a select returns successfully, the ProcMgr_waitForMultipleEvents
call unregisters all the events that were passed in rather than only the
one which got triggered. This does not rely on the application to close
all the syslink drivers between sucessive use cases to prevent the memory
leak. This is the common use with applications currently as the de-init
IPC stack after each video/image usecase. The kernel closes the the syslink
drivers and prevents a memory leak in this scenario.

Signed-off-by: Wajahat Khan <w-khan@ti.com>