SysLink: Utils - Remove system gate from List module
The List module (utils/List.c) is using the system gate
for protection along with an instance specific gate supplied
during creation.  There is no need for both, so the system
gate is being removed.  This prevents possible freezes if
the instance gate is set to the system gate.  Lists will
not use a gate if one is not assigned during creation.

Signed-off-by: Cris Jansson <cjansson@ti.com>