SysLink: Daemons - Strip the commandline till the last slash
This patch fixes an issue with detecting multiple instances of
the SysLink Daemon if they are launched from different paths.

The initial assumption was, daemon will always be executed from
the same directory. But in reality this may not be the case. It
can be executed from any directory and in that case the current
implementation would fail the detection test. For eg., first time
if the daemon is executed from /system/bin folder, the entry in
the procfs will be ./syslink_daemon.out, and second time if it is
executed from /system patg, it will be ./bin/syslink_daemon.out,
and the current implementation considers it as a different
executable. This patch fixes this issue by stripping the command
line upto the last '/' and then comparing the rest.

A KW fix, sprintf replaced with snprintf for boundary checking,
is also added.

Signed-off-by: Subin K G <subin.kg@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>