SysLink: Daemons - Add an option to not run as a daemon
A new option "-f" is added to command the SysLink daemon
to run as a foreground process. This change is required if
the syslink_daemon.out is running as a init.rc service and
need to be restarted automatically when it crashed or was
killed.

Example Usage:
./syslink_daemon.out -f <SysM3image> <AppM3image>
This will force the daemon to run as foreground process.

Signed-off-by: Subin K G <subin.kg@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>