SysLink: ProcMgr - Retrieve API status for Start & Stop
The status for the start and stop remoteproc ioctls is not
retrieved as there are no arguments being passed to the ioctls.
The status is always implied to be a SUCCESS, irrespective of
the actual return value. This is fixed by introducing a remoteproc
specific ioctl command arguments.

The patch also fixes an issue with returning the apiStatus for all
the ProcMgr ioctl calls.

Signed-off-by: Suman Anna <s-anna@ti.com>