SysLink: Samples - Start notify samples from event 6
Update the Notify sample so that the sample starts the registration
from event 6 instead of 5. This is needed as event 5 is used to
implement the HDMI workaround. Retaining the value at 5 conflicts
with the event used by the HDMI driver.

The code comments and traces have also been corrected accordingly.

Signed-off-by: Bryan Honza <honza@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>